backend/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# ML/AI Libraries
openai-whisper==20231117
deepface==0.0.79
tf-keras==2.15.0
tensorflow==2.15.0

# Media Processing
ffmpeg-python==0.2.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
